avoid hardcoding the cmake prefix path
I'd prefer to illustrate the practice of leaving it explicitly
to the user's build configuration so they can decide
where to point it at. Hardcoded like this it always ends up
being a suprise when the install location is not there (and there
is no agreed upon convention that insists that it should be there).