Workaround bug in Ubuntu libvtk6-dev package.
Workaround for https://bugs.launchpad.net/ubuntu/+source/vtk6/+bug/1573234 which
affects our PCL example when not using the backported ROS libvtk6-dev package.
Since this library is not needed by the example, the workaround does not need to
check the installed package version.

In order to compile without trying to link the missing vtkproj4, the package
libproj-dev needs to be installed but I did not need to do any additional
configuration to incorporate it into the build. I haven't explored why.